requirements.txt
Flask==2.0.3
flask-cors==3.0.10
yfinance==0.1.63
firebase-admin==4.5.3
openai==1.12.0
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.3
google-cloud-firestore==2.3.4
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.0.2
